requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.28.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.0
python-dotenv>=1.0.0
gradio>=4.0.0
playwright>=1.40.0
pyautogui>=0.9.54
Pillow>=10.0.0
